[FIX] pos_sale: prevent setting lot when user cancels lot selection
Before this commit, when loading a sale order containing an order line
tracked by lot, the system prompted the user to select a lot. However,
even if the user canceled the selection, the lot was still added to the
order line.

After this commit, the lot will no longer be set if the user cancels
the selection.

opw-5162487

closes odoo/odoo#233859

X-original-commit: e0d95a1b1158589e36598e7a56db1004849f8ead

+++ addons/pos_sale/static/src/app/services/pos_store.js
@@ -182,7 +182,11 @@ patch(PosStore.prototype, {
             }
 
             // Order line can only hold one lot, so we need to split the line if there are multiple lots
-            if (line.product_id.tracking == "lot" && converted_line.lot_names.length > 0) {
+            if (
+                line.product_id.tracking == "lot" &&
+                converted_line.lot_names.length > 0 &&
+                useLoadedLots
+            ) {
                 newLine.delete();
                 for (const lot of converted_line.lot_names) {
                     const splitted_line = this.models["pos.order.line"].create({
